move entity to core
