+++ backend/src/core/entities/friend_request.entity.ts
